langchain/auto/requirements.txt
# LangChain and Anthropic
langchain>=0.3.0
langchain-anthropic>=0.3.0
anthropic>=0.40.0

# OpenTelemetry Auto-Instrumentation (matching versions)
opentelemetry-api==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-distro==0.60b1
opentelemetry-exporter-otlp==1.39.1
opentelemetry-instrumentation==0.60b1
opentelemetry-instrumentation-langchain==0.51.0
opentelemetry-semantic-conventions==0.60b1
opentelemetry-semantic-conventions-ai==0.4.13

# Utilities
python-dotenv>=1.0.0
